Collect servers in betty.servers
