+++ betty/commands/docs.py
@@ -3,11 +3,11 @@ from __future__ import annotations  # noqa: D100
 import asyncio
 from typing import TYPE_CHECKING, Self, final, override
 
-from betty import documentation
 from betty.app import App
 from betty.console.command import Command, CommandDefinition, CommandFunction
 from betty.factory import Manufacturable
 from betty.locale.localizable.gettext import _
+from betty.servers import documentation
 
 if TYPE_CHECKING:
     import argparse
