+++ betty/documentation.py
@@ -6,18 +6,14 @@ from __future__ import annotations
 
 import multiprocessing
 from asyncio import to_thread
-from contextlib import AsyncExitStack
 from shutil import copytree
-from typing import TYPE_CHECKING, final, override
+from typing import TYPE_CHECKING
 
 from sphinx.application import Sphinx
 from sphinx.ext.autodoc import MethodDocumenter
 
 from betty.dirs import root_directory
 from betty.exception import HumanFacingException
-from betty.pathlib import StrPath, resolve_path
-from betty.server import Server, ServerNotStarted
-from betty.servers import builtin
 from betty.user import User, Verbosity
 
 if TYPE_CHECKING:
@@ -63,36 +59,3 @@ async def _build(output_directory: Path, cache_directory: Path, *, user: User) -
         MethodDocumenter.can_document_member = original_can_document_member  # ty:ignore[invalid-assignment]
     if sphinx_app.statuscode != 0:
         raise HumanFacingException("Sphinx failed.")
-
-
-@final
-class DocumentationServer(Server):
-    """
-    Serve the documentation site.
-    """
-
-    def __init__(self, cache_directory: StrPath, *, user: User):
-        super().__init__(user=user)
-        self._cache_directory = resolve_path(cache_directory)
-        self._server: Server | None = None
-        self._exit_stack = AsyncExitStack()
-
-    @override
-    @property
-    def public_url(self) -> str:
-        if self._server is not None:
-            return self._server.public_url
-        raise ServerNotStarted
-
-    @override
-    async def start(self) -> None:
-        www_directory = self._cache_directory / "www"
-        await _ensure_www_directory(
-            www_directory, self._cache_directory / "cache", user=self._user
-        )
-        self._server = builtin.BuiltinServer(www_directory, user=self._user)
-        await self._exit_stack.enter_async_context(self._server)
-
-    @override
-    async def stop(self) -> None:
-        await self._exit_stack.aclose()
