+++ betty/tests/commands/test_docs.py
@@ -2,13 +2,13 @@ from pytest_mock import MockerFixture
 
 from betty.app import App
 from betty.console import SystemExitCode
+from betty.servers.no_op import NoOpServer
 from betty.test_utils.console import run
-from betty.test_utils.server import NoOpServer
 
 
 class TestDocs:
     async def test_configure(self, mocker: MockerFixture, isolated_app: App) -> None:
-        mocker.patch("betty.documentation.DocumentationServer", new=NoOpServer)
+        mocker.patch("betty.servers.documentation.DocumentationServer", new=NoOpServer)
         mocker.patch(
             "betty.commands.docs:Docs._wait_forever",
             side_effect=KeyboardInterrupt,
