+++ betty/tests/test_documentation.py
@@ -5,33 +5,14 @@ from collections.abc import Iterator
 from pathlib import Path
 
 import pytest
-import requests
-from pytest_mock import MockerFixture
-from requests import Response
 from sphinx.errors import ExtensionError
 from sphinx.util import import_object
 
 from betty.app import App
 from betty.console.command import CommandDefinition
 from betty.dirs import root_directory
-from betty.documentation import DocumentationServer
-from betty.functools import Do
 from betty.locale.localize import default_localizer
 from betty.test_utils.documentation import PluginDocumentationTestBase
-from betty.test_utils.user import StaticUser
-
-
-class TestDocumentationServer:
-    @pytest.mark.order(0)
-    async def test(self, mocker: MockerFixture, tmp_path: Path) -> None:
-        mocker.patch("webbrowser.open_new_tab")
-        async with DocumentationServer(tmp_path, user=StaticUser()) as server:
-
-            def _assert_response(response: Response) -> None:
-                assert response.status_code == 200
-                assert "Betty" in response.content.decode("utf-8")
-
-            await Do(requests.get, server.public_url).until(_assert_response)
 
 
 class TestDocumentation:
